api actions created

+++ database/migrations/2022_10_30_203501_create_cars_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('cars', function (Blueprint $table) {
             $table->id();
             $table->string('model');
-            $table->unsignedTinyInteger('year');
+            $table->unsignedSmallInteger('year');
             $table->enum('color', ['white', 'black', 'blue', 'red']);
             $table->timestamps();
         });
